refactor: move settings module to obsidian
