+++ src/main.ts
@@ -3,7 +3,7 @@ import { Reminder, Reminders } from 'model/reminder';
 import { DATE_TIME_FORMATTER } from 'model/time';
 import { App, Plugin, PluginManifest } from 'obsidian';
 import { monkeyPatchConsole } from 'obsidian/obsidian-hack/obsidian-debug-mobile';
-import { SETTINGS } from 'settings';
+import { SETTINGS } from 'obsidian/settings';
 import { ReminderListItemViewProxy } from 'ui/reminder-list';
 import { registerCommands } from 'obsidian/commands';
 import { ReminderPluginUI } from 'obsidian/ui';
