+++ src/obsidian/ui.ts
@@ -4,7 +4,7 @@ import type { DateTime } from 'model/time';
 import type { ReminderListItemViewProxy } from 'ui/reminder-list';
 import type { Reminder } from 'model/reminder';
 import { MarkdownView, Platform, TFile, WorkspaceLeaf } from 'obsidian';
-import { ReminderSettingTab, SETTINGS } from 'settings';
+import { ReminderSettingTab, SETTINGS } from 'obsidian/settings';
 import { AutoCompletableEditor, AutoComplete } from 'ui/autocomplete';
 import { DateTimeChooserView } from 'ui/datetime-chooser';
 import { buildCodeMirrorPlugin } from 'ui/editor-extension';
